+++ src/Queries.Core/Builders/Fluent/IJoinQuery.cs
@@ -2,7 +2,11 @@
 
 namespace Queries.Core.Builders.Fluent
 {
-    public interface IJoinQuery<T> : IInsertable
+    /// <summary>
+    /// Allows to append a <see cref="IWhereClause"/> to the query being built
+    /// </summary>
+    /// <typeparam name="T">Type of the query under construction</typeparam>
+    public interface IJoinQuery<T> : IInsertable, IBuild<T>
     {
         IWhereQuery<T> Where(IWhereClause clause);
     }